requirements.txt
openai>=1.0.0
httpx>=0.24.0
Pillow>=9.0.0
torch>=1.9.0
numpy>=1.21.0
httpx[socks]
